+++ src/main/scala/Core/IS/AiqEntry.scala
@@ -1,4 +1,4 @@
-package Core.RS
+package Core.IS
 
 import chisel3._
 import chisel3.util._
