+++ src/test/scala/Core/IS/AiqEntryMain.scala
@@ -1,12 +1,12 @@
-package Core.RS
+package Core.IS
 
 import chisel3.stage.{ChiselGeneratorAnnotation, ChiselStage}
 
-object ArithInstQueueMain extends App {
+object AiqEntryMain extends App {
   (new ChiselStage).execute(
     args,
     Seq(
-      ChiselGeneratorAnnotation(() => new ArithInstQueue)
+      ChiselGeneratorAnnotation(() => new AiqEntry)
     )
   )
 }
