+++ src/test/scala/Core/IS/DepRegEntryMain.scala
@@ -1,12 +1,12 @@
-package Core.RS
+package Core.IS
 
 import chisel3.stage.{ChiselGeneratorAnnotation, ChiselStage}
 
-object AiqEntryMain extends App {
+object DepRegEntryMain extends App {
   (new ChiselStage).execute(
     args,
     Seq(
-      ChiselGeneratorAnnotation(() => new AiqEntry)
+      ChiselGeneratorAnnotation(() => new DepRegEntry)
     )
   )
 }
